refactor(server): move live task models beside workflow runtime
